Move basic items from ext.commands.Bot to Client
